Move all files related to the search functionality into a new components/search folder
